+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Delta Chat Android Changelog
 
-## Unreleased
+## 2.51.0
+2026-05
 
 * Better incoming call system integration
 * Calls are not experimental anymore and don't need to be manually enabled